Fixed return type CardIdGenerator Added Player class

+++ src/bg/sofia/uni/fmi/mjt/id/CardIdGenerator.java
@@ -1,7 +1,7 @@
 package bg.sofia.uni.fmi.mjt.id;
 
 public class CardIdGenerator implements IdGenerator {
-    long id = 1;
+    int id = 1;
 
     @Override
     public int getId() {
